Obtain low resolution tasksnapshots from disk to improve the smoothness of recent tasks' sliding
Firstly, add a property switch named GET_SNAPSHOTS_FROM_DISK, with a default value of false. Secondly, when obtaining tasksnapshots from the disk, set the value of GET_SNAPSHOTS_FROM_DISK to true. Thirdly, when the value of GET_SNAPSHOTS_FROM_DISK is true, the parameter isLowResolution of the getSnapshot method passes in a value of true, which means taking a low resolution snapshot.

When entry is not null and entry.snapshot is not null, only the following four situations will obtain tasksnapshots from the cache:
1.When GET_SNAPSHOTS_FROM_DISK is false, directly obtain snapshots from cache.
2.When isLowResolution is false, obtain snapshots from cache.
3.When obtaining low resolution snapshots and the snapshots stored in the cache are low resolution, obtain snapshots from the cache.
4.When restoreFromDisk is false, obtain snapshot from cache.

bug:293213295

Change-Id: I3b6610231a20d97aa173e6144dae2c88b1338b82
Signed-off-by: tingting <tingtingteng0@gmail.com>